Moving static functions into crypto and using ZERO_TO_DISABLE_DEFER_CLEANUP
