+++ crypto/s2n_openssl_x509.c
@@ -13,19 +13,12 @@
  * permissions and limitations under the License.
  */
 
-#pragma once
+#include "crypto/s2n_openssl_x509.h"
 
-#include <openssl/aes.h>
-#include <openssl/rc4.h>
-#include <openssl/des.h>
-#include <openssl/rsa.h>
-#include <openssl/dh.h>
-
-/* OPENSSL_free is defined within <openssl/crypto.h> for OpenSSL Libcrypto
- * and within <openssl/mem.h> for AWS_LC */
-#include <openssl/crypto.h>
-#if defined(OPENSSL_IS_AWSLC)
-#include <openssl/mem.h>
-#endif
-
-int s2n_openssl_free(uint8_t** data);
+int s2n_sk_X509_pop_free(STACK_OF(X509) **cert_chain)
+{
+    if (*cert_chain != NULL) {
+        sk_X509_pop_free(*cert_chain, X509_free);
+    }
+    return 0;
+}
