+++ crypto/s2n_openssl_x509.h
@@ -20,3 +20,5 @@
 #include "utils/s2n_safety.h"
 
 DEFINE_POINTER_CLEANUP_FUNC(X509*, X509_free);
+
+int s2n_sk_X509_pop_free(STACK_OF(X509) **cert_chain);
